Created linkshellMgr class and added basic linkshell caching

+++ src/servers/Server_Zone/Linkshell/Linkshell.h
@@ -11,49 +11,49 @@ class Linkshell
 {
 private:
    /*! unique ID of the linkshell */
-   uint32_t m_linkshellId;
+   uint64_t m_linkshellId;
    /*! ID of the master character */
-   uint32_t m_masterCharacterId;
+   uint64_t m_masterCharacterId;
    /*! ID list of all linkshell members */
-   std::set< uint32_t > m_memberIds;
+   std::set< uint64_t > m_memberIds;
    /*! Name of the linkshell */
    std::string m_name;
    /*! List of member IDs with leader rank */
-   std::set< uint32_t > m_leaderIds;
+   std::set< uint64_t > m_leaderIds;
    /*! list of IDs of pending character invites */
-   std::set< uint32_t > m_inviteIds;
+   std::set< uint64_t > m_inviteIds;
 
 public:
-   Linkshell( uint32_t id,
+   Linkshell( uint64_t id,
               const std::string& name,
-              uint32_t masterId,
-              const std::set< uint32_t >& members,
-              const std::set< uint32_t >& leaders,
-              const std::set< uint32_t >& invites );
+              uint64_t masterId,
+              const std::set< uint64_t >& members,
+              const std::set< uint64_t >& leaders,
+              const std::set< uint64_t >& invites );
 
-   uint32_t getId() const;
+   uint64_t getId() const;
 
    const std::string& getName() const;
 
-   uint32_t getMasterId() const;
+   uint64_t getMasterId() const;
 
-   const std::set< uint32_t >& getMemberIdList() const;
-   std::set< uint32_t >& getMemberIdList();
+   const std::set< uint64_t >& getMemberIdList() const;
+   std::set< uint64_t >& getMemberIdList();
 
-   const std::set< uint32_t >& getLeaderIdList() const;
-   std::set< uint32_t >& getLeaderIdList();
+   const std::set< uint64_t >& getLeaderIdList() const;
+   std::set< uint64_t >& getLeaderIdList();
 
-   const std::set< uint32_t >& getInviteIdList() const;
-   std::set< uint32_t >& getInviteIdList();
+   const std::set< uint64_t >& getInviteIdList() const;
+   std::set< uint64_t >& getInviteIdList();
 
-   void addMember( uint32_t memberId );
-   void removeMember( uint32_t memberId );
+   void addMember( uint64_t memberId );
+   void removeMember( uint64_t memberId );
 
-   void addLeader( uint32_t memberId );
-   void removeLeader( uint32_t memberId );
+   void addLeader( uint64_t memberId );
+   void removeLeader( uint64_t memberId );
 
-   void addInvite( uint32_t memberId );
-   void removeInvite( uint32_t memberId );
+   void addInvite( uint64_t memberId );
+   void removeInvite( uint64_t memberId );
 
 
 };
